code/requirements.txt
openai
gTTS
playsound
unidecode
zmq
opencv-python
Pillow
google-cloud-texttospeech
PyQt6